+++ dockstore-webservice/src/main/java/io/dockstore/webservice/jdbi/AbstractDockstoreDAO.java
@@ -1,0 +1,35 @@
+/*
+ *    Copyright 2018 OICR
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package io.dockstore.webservice.jdbi;
+
+import io.dropwizard.hibernate.AbstractDAO;
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+
+/**
+ * Standardizes additional methods expected of Dockstore DAOs.
+ */
+public abstract class AbstractDockstoreDAO<T> extends AbstractDAO<T> {
+
+    public AbstractDockstoreDAO(SessionFactory factory) {
+        super(factory);
+    }
+
+    public void evict(Object entry) {
+        Session session = currentSession();
+        session.evict(entry);
+    }
+}
