+++ dockstore-webservice/src/main/java/io/dockstore/webservice/jdbi/UserDAO.java
@@ -19,15 +19,13 @@ package io.dockstore.webservice.jdbi;
 import java.util.List;
 
 import io.dockstore.webservice.core.User;
-import io.dropwizard.hibernate.AbstractDAO;
 import org.hibernate.Query;
-import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 
 /**
  * @author xliu
  */
-public class UserDAO extends AbstractDAO<User> {
+public class UserDAO extends AbstractDockstoreDAO<User> {
     public UserDAO(SessionFactory factory) {
         super(factory);
     }
@@ -43,11 +41,6 @@ public class UserDAO extends AbstractDAO<User> {
     public void clearCache() {
         currentSession().flush();
         currentSession().clear();
-    }
-
-    public void evict(Object entry) {
-        Session session = currentSession();
-        session.evict(entry);
     }
 
     public List<User> findAll() {
