+++ dockstore-webservice/src/main/resources/swagger.yaml
@@ -4,7 +4,7 @@ info:
   description: "This describes the dockstore API, a webservice that manages pairs\
     \ of Docker images and associated metadata such as CWL documents and Dockerfiles\
     \ used to build those images"
-  version: "1.3.0"
+  version: "1.4.4"
   title: "Dockstore API"
   termsOfService: ""
   contact:
@@ -12,8 +12,8 @@ info:
     url: "https://github.com/ga4gh/dockstore"
     email: "theglobalalliance@genomicsandhealth.org"
   license:
-    name: " GNU Lesser General Public License"
-    url: "https://www.gnu.org/licenses/lgpl-3.0.en.html"
+    name: "Apache License Version 2.0"
+    url: "https://github.com/ga4gh/dockstore/blob/develop/LICENSE"
 host: ""
 basePath: "/"
 tags:
